LayoutTests/inspector/css-syntax-highlight.html

<html>
<head>
<script src="evaluate-in-frontend.js"></script>
<script>

function doit()
{
    function callback(result)
    {
        for (var i = 0; i < result.length; ++i)
            output(result[i]);
        notifyDone();
    }
    evaluateInWebInspector("doitAndDump", callback);
}

</script>
</head>

<body onload="onload()">
<p>
Tests that CSSSourceSyntaxHighlighter detects the tokens.
</p>

<div id="frontend-script" style="display:none">
function doitAndDump(testController) {
    testController.waitUntilDone();
    setTimeout(function() {
        var result = [];
        
        result.push(dumpSyntaxHighlight("a[href='/']"));
        result.push(dumpSyntaxHighlight("#content > a:hover"));
        result.push(dumpSyntaxHighlight("@import url(style.css);"));
        result.push(dumpSyntaxHighlight("@import url(\"style.css\") projection, tv;"));
        result.push(dumpSyntaxHighlight("@media screen { body { color: red; } }"));
        result.push(dumpSyntaxHighlight("@font-face { font-family: MyHelvetica; }"));
        result.push(dumpSyntaxHighlight("p { color: color; red: red; }"));
        result.push(dumpSyntaxHighlight("p { color: red !important; }"));
        
        testController.notifyDone(result);
        
    }, 0);
}
function dumpSyntaxHighlight(str) {
    var node = document.createElement("span");
    node.textContent = str;
    var javascriptSyntaxHighlighter = new WebInspector.CSSSourceSyntaxHighlighter(null, null);
    javascriptSyntaxHighlighter.syntaxHighlightNode(node);
    var node_parts = [];
    for (var i = 0; i < node.childNodes.length; i++) {
        if (node.childNodes[i].getAttribute) {
            node_parts.push(node.childNodes[i].getAttribute("class"));
        } else {
            node_parts.push("*");
        }
    }
    return node_parts.join(",");
}
</div>

<div id="output">
</div>

</body>
</html>
